Avoid hooks
